+++ src/03 - Domain/Domain/Dtos/Finance/DespesasPorMesDto.cs
@@ -1,5 +1,4 @@
 ﻿namespace Domain.Dtos.Finance
 {
-    public record DespesasPorMesDto(string Mes, decimal TotalDespesas);
-
+    public record DespesasPorMesDto(string Mes, double TotalDespesas);
 }
