+++ src/03 - Domain/Domain/Dtos/Finance/DespesasTotalPorCategoria.cs
@@ -1,5 +1,4 @@
 ﻿namespace Domain.Dtos.Finance
 {
-    public record DespesasTotalPorCategoria(string Categoria, decimal Total);
-
+    public record DespesasTotalPorCategoria(string Categoria, double Total);
 }
